Move system setup to system plugin
